BM_core/requirements.txt
celery==5.2.3
charset-normalizer==2.0.12
Django==4.0.1
django-celery-results==2.4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
plaid-python==7.4.0
redis==4.3.3
requests==2.28.0
